third_party/blink/web_tests/platform/mac/external/wpt/editing/run/bold_2001-3000-expected.txt

This is a testharness.js-based test.
Found 180 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["bold",""]] "<span style=\\"font-weight: 900\\">foo[barbaz</span>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:900\\">foo</span>barbaz" but got "<span style=\\"font-weight:900\\">foobarbaz</span>"
[FAIL] [["bold",""]] "<span style=\\"font-weight: 900\\">foo[barbaz</span>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span style=\\"font-weight: 900\\">foo[barbaz</span>}" queryCommandState("bold") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["bold",""]] "<span style=\\"font-weight: 900\\">foo[barbaz</span>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>foo[bar]baz</h3>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>foo[bar]baz</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>foo[bar]baz</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>foo[bar]baz</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>foo[bar]baz</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foobar]baz</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foobar]baz</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foobar]baz</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foobar]baz</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>foo[barbaz</h3>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>foo[barbaz</h3>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>foo[barbaz</h3>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>foo[barbaz</h3>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>[foobarbaz]</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>[foobarbaz]</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>[foobarbaz]</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>[foobarbaz]</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foobarbaz]</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foobarbaz]</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foobarbaz]</h3>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foobarbaz]</h3>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>[foobarbaz</h3>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<h3>[foobarbaz</h3>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>[foobarbaz</h3>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<h3>[foobarbaz</h3>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foobarbaz</h3>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foobarbaz</h3>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foobarbaz</h3>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foobarbaz</h3>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">bar<b>[baz]</b>quz</span>qoz</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo<span style=\\"font-weight:normal\\">barbazquz</span>qoz</b>" but got "<span style=\\"font-weight:bold\\">foo</span>barbazquz<span style=\\"font-weight:bold\\">qoz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">bar<b>[baz]</b>quz</span>qoz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">bar<b>[baz]</b>quz</span>qoz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">bar<b>[baz]</b>quz</span>qoz</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo<span style=\\"font-weight:normal\\">barbazquz</span>qoz</b>" but got "<b>foo</b>barbazquz<b>qoz</b>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">bar<b>[baz]</b>quz</span>qoz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">bar<b>[baz]</b>quz</span>qoz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">[bar]</span>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">[bar]</span>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">[bar]</span>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo<span style=\\"font-weight: normal\\">[bar]</span>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<b>foo</b> <b>bar</b>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:bold\\">foo bar</span>" but got "foo bar"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandState("bold") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<b>foo</b> <b>bar</b>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo bar</b>" but got "foo bar"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandState("bold") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["bold",""]] "{<b>foo</b> <b>bar</b>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foo</h3><b>bar</b>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<h3>foo</h3><b>bar</b>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foo</h3><b>bar</b>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<h3>foo</h3><b>bar</b>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<i><b>foo</b></i>[bar]<i><b>baz</b></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b><i>foo</i>bar<i>baz</i></b>" but got "<i><b>foo</b></i><span style=\\"font-weight:bold\\">bar</span><i><b>baz</b></i>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<i><b>foo</b></i>[bar]<i><b>baz</b></i>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<i><b>foo</b></i>[bar]<i><b>baz</b></i>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<i><b>foo</b></i>[bar]<i><b>baz</b></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b><i>foo</i>bar<i>baz</i></b>" but got "<i><b>foo</b></i><b>bar</b><i><b>baz</b></i>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<i><b>foo</b></i>[bar]<i><b>baz</b></i>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<i><b>foo</b></i>[bar]<i><b>baz</b></i>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<i><b>foo</b></i>[bar]<b>baz</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b><i>foo</i>barbaz</b>" but got "<i><b>foo</b></i><span style=\\"font-weight:bold\\">bar</span><b>baz</b>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<i><b>foo</b></i>[bar]<b>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<i><b>foo</b></i>[bar]<b>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<i><b>foo</b></i>[bar]<b>baz</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b><i>foo</i>barbaz</b>" but got "<i><b>foo</b></i><b>barbaz</b>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<i><b>foo</b></i>[bar]<b>baz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<i><b>foo</b></i>[bar]<b>baz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo</b>[bar]<i><b>baz</b></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foobar<i>baz</i></b>" but got "<b>foo</b><span style=\\"font-weight:bold\\">bar</span><i><b>baz</b></i>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo</b>[bar]<i><b>baz</b></i>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>foo</b>[bar]<i><b>baz</b></i>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo</b>[bar]<i><b>baz</b></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foobar<i>baz</i></b>" but got "<b>foobar</b><i><b>baz</b></i>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo</b>[bar]<i><b>baz</b></i>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>foo</b>[bar]<i><b>baz</b></i>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<font color=blue face=monospace><b>foo</b></font>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b><font color=\\"blue\\" face=\\"monospace\\">foo</font>bar</b>" but got "<font color=\\"blue\\" face=\\"monospace\\"><b>foo</b></font><b>bar</b>"
[FAIL] [["bold",""]] "<font color=blue face=monospace><b>foo</b></font>[bar]" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "<font color=blue face=monospace><b>foo</b></font>[bar]" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: normal\\"><b>{bar}</b></span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: normal\\"><b>{bar}</b></span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: normal\\"><b>{bar}</b></span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: normal\\"><b>{bar}</b></span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "[foo<span class=notbold>bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:bold\\">foo<span class=\\"notbold\\"><span style=\\"font-weight:bold\\">bar</span></span>baz</span>" but got "<span style=\\"font-weight:bold\\">foo<span class=\\"notbold\\">bar</span>baz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "[foo<span class=notbold>bar</span>baz]" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "[foo<span class=notbold>bar</span>baz]" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "[foo<span class=notbold>bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo<span class=\\"notbold\\"><b>bar</b></span>baz</b>" but got "<b>foo<span class=\\"notbold\\">bar</span>baz</b>"
[FAIL] [["stylewithcss","false"],["bold",""]] "[foo<span class=notbold>bar</span>baz]" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "[foo<span class=notbold>bar</span>baz]" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><span class=notbold>[foo]</span></b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span class=\\"notbold\\"><span style=\\"font-weight:bold\\">foo</span></span>" but got "<span class=\\"notbold\\" style=\\"font-weight:bold\\">foo</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><span class=notbold>[foo]</span></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><span class=notbold>[foo]</span></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b><span class=notbold>[foo]</span></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b><span class=notbold>[foo]</span></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><span class=notbold>foo[bar]baz</span></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><span class=notbold>foo[bar]baz</span></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b><span class=notbold>foo[bar]baz</span></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b><span class=notbold>foo[bar]baz</span></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-weight:bold\\">foo</span>bar<span style=\\"font-weight:bold\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-weight:bold\\">foo</span>bar<span style=\\"font-weight:bold\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b>bar<b>baz</b></p>" but got "<p style=\\"\\"><span style=\\"font-weight:bold\\">foo</span>bar<b>baz</b></p>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "fo[o<b>b]ar</b>baz" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""]] "fo[o<b>b]ar</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "fo[o<b>b]ar</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<b>ba[r</b>b]az" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>barb</b>az" but got "foo<b>ba</b>rbaz"
[FAIL] [["bold",""]] "foo<b>ba[r</b>b]az" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""]] "foo<b>ba[r</b>b]az" queryCommandState("bold") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["bold",""]] "foo<b>ba[r</b>b]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo<b>ba[r</b>b]az" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""]] "foo<b>ba[r</b>b]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "fo[o<b>bar</b>b]az" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["bold",""]] "fo[o<b>bar</b>b]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "fo[o<b>bar</b>b]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "fo[o<b>bar</b>b]az" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["bold",""]] "fo[o<b>bar</b>b]az" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "fo[o<b>bar</b>b]az" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<b>b]ar</b>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foob<span style=\\"font-weight:bold\\">ar</span>baz" but got "foob<b>ar</b>baz"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<b>b]ar</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<b>b]ar</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<b>b]ar</b>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<b>b]ar</b>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<b>ba[r</b>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-weight:bold\\">ba</span>rbaz" but got "foo<b>ba</b>rbaz"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<b>ba[r</b>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<b>ba[r</b>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<b>ba[r</b>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<b>ba[r</b>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo{<b>bar</b>}baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo{<b>bar</b>}baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo{<b>bar</b>}baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo{<b>bar</b>}baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "fo[o<span style=font-weight:bold>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<span style=\\"font-weight:bold\\">obar</span>baz" but got "fo<b>ob</b><span style=\\"font-weight:bold\\">ar</span>baz"
[FAIL] [["bold",""]] "fo[o<span style=font-weight:bold>b]ar</span>baz" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""]] "fo[o<span style=font-weight:bold>b]ar</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "fo[o<span style=font-weight:bold>b]ar</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span style=font-weight:800>fo[o</span><span style=font-weight:900>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "<span style=font-weight:800>fo[o</span><span style=font-weight:900>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:700>fo[o</span><span style=font-weight:800>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:700>fo[o</span><span style=font-weight:800>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:700>fo[o</span><span style=font-weight:800>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>fo</b>ob<span style=\\"font-weight:800\\">ar</span>" but got "<span style=\\"font-weight:700\\">fo</span>ob<span style=\\"font-weight:800\\">ar</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:700>fo[o</span><span style=font-weight:800>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:700>fo[o</span><span style=font-weight:800>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:600>fo[o</span><span style=font-weight:700>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:600>fo[o</span><span style=font-weight:700>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:600>fo[o</span><span style=font-weight:700>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:600\\">fo</span>ob<b>ar</b>" but got "<span style=\\"font-weight:600\\">fo</span>ob<span style=\\"font-weight:700\\">ar</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:600>fo[o</span><span style=font-weight:700>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:600>fo[o</span><span style=font-weight:700>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:500\\">fo<span style=\\"font-weight:bold\\">o</span></span><span style=\\"font-weight:600\\"><span style=\\"font-weight:bold\\">b</span>ar</span>" but got "<span style=\\"font-weight:500\\">fo</span><span style=\\"font-weight:bold\\">ob</span><span style=\\"font-weight:600\\">ar</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:500\\">fo<b>o</b></span><span style=\\"font-weight:600\\"><b>b</b>ar</span>" but got "<span style=\\"font-weight:500\\">fo</span><b>ob</b><span style=\\"font-weight:600\\">ar</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" queryCommandIndeterm("bold") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:500>fo[o</span><span style=font-weight:600>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:400>fo[o</span><span style=font-weight:500>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\">fo<span style=\\"font-weight:bold\\">o</span></span><span style=\\"font-weight:500\\"><span style=\\"font-weight:bold\\">b</span>ar</span>" but got "<span style=\\"font-weight:400\\">fo</span><span style=\\"font-weight:bold\\">ob</span><span style=\\"font-weight:500\\">ar</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:400>fo[o</span><span style=font-weight:500>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:400>fo[o</span><span style=font-weight:500>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:400>fo[o</span><span style=font-weight:500>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\">fo<b>o</b></span><span style=\\"font-weight:500\\"><b>b</b>ar</span>" but got "<span style=\\"font-weight:400\\">fo</span><b>ob</b><span style=\\"font-weight:500\\">ar</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:400>fo[o</span><span style=font-weight:500>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:400>fo[o</span><span style=font-weight:500>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:300>fo[o</span><span style=font-weight:400>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:300\\">fo<span style=\\"font-weight:bold\\">o</span></span><span style=\\"font-weight:400\\"><span style=\\"font-weight:bold\\">b</span>ar</span>" but got "<span style=\\"font-weight:300\\">fo</span><span style=\\"font-weight:bold\\">ob</span><span style=\\"font-weight:400\\">ar</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:300>fo[o</span><span style=font-weight:400>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:300>fo[o</span><span style=font-weight:400>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:300>fo[o</span><span style=font-weight:400>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:300\\">fo<b>o</b></span><span style=\\"font-weight:400\\"><b>b</b>ar</span>" but got "<span style=\\"font-weight:300\\">fo</span><b>ob</b><span style=\\"font-weight:400\\">ar</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:300>fo[o</span><span style=font-weight:400>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:300>fo[o</span><span style=font-weight:400>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:200>fo[o</span><span style=font-weight:300>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:200\\">fo<span style=\\"font-weight:bold\\">o</span></span><span style=\\"font-weight:300\\"><span style=\\"font-weight:bold\\">b</span>ar</span>" but got "<span style=\\"font-weight:200\\">fo</span><span style=\\"font-weight:bold\\">ob</span><span style=\\"font-weight:300\\">ar</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:200>fo[o</span><span style=font-weight:300>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:200>fo[o</span><span style=font-weight:300>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:200>fo[o</span><span style=font-weight:300>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:200\\">fo<b>o</b></span><span style=\\"font-weight:300\\"><b>b</b>ar</span>" but got "<span style=\\"font-weight:200\\">fo</span><b>ob</b><span style=\\"font-weight:300\\">ar</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:200>fo[o</span><span style=font-weight:300>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:200>fo[o</span><span style=font-weight:300>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\">fo<span style=\\"font-weight:bold\\">o</span></span><span style=\\"font-weight:200\\"><span style=\\"font-weight:bold\\">b</span>ar</span>" but got "<span style=\\"font-weight:100\\">fo</span><span style=\\"font-weight:bold\\">ob</span><span style=\\"font-weight:200\\">ar</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=font-weight:100>fo[o</span><span style=font-weight:200>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\">fo<b>o</b></span><span style=\\"font-weight:200\\"><b>b</b>ar</span>" but got "<span style=\\"font-weight:100\\">fo</span><b>ob</b><span style=\\"font-weight:200\\">ar</span>"
Harness: the test ran to completion.
